+++ src/085-the-utils-folder/221-type-predicates.solution.ts
@@ -1,16 +1,74 @@
 import { Equal, Expect } from "@total-typescript/helpers";
-import { expect, it } from "vitest";
+import { describe, expect, it } from "vitest";
 
-const isString = (input: unknown): input is string => {
-  return typeof input === "string";
+const hasDataAndId = (
+  value: unknown,
+): value is {
+  data: {
+    id: string;
+  };
+} => {
+  return (
+    typeof value === "object" &&
+    value !== null &&
+    "data" in value &&
+    typeof value.data === "object" &&
+    value.data !== null &&
+    "id" in value.data &&
+    typeof value.data.id === "string"
+  );
 };
 
-it("Should be able to be passed to .filter and work", () => {
-  const mixedArray = [1, "hello", [], {}];
+const parseValue = (value: unknown) => {
+  if (hasDataAndId(value)) {
+    return value.data.id;
+  }
 
-  const stringsOnly = mixedArray.filter(isString);
+  throw new Error("Parsing error!");
+};
 
-  type test1 = Expect<Equal<typeof stringsOnly, string[]>>;
+const parseValueAgain = (value: unknown) => {
+  if (hasDataAndId(value)) {
+    return value.data.id;
+  }
 
-  expect(stringsOnly).toEqual(["hello"]);
+  throw new Error("Parsing error!");
+};
+
+describe("parseValue", () => {
+  it("Should handle a { data: { id: string } }", () => {
+    const result = parseValue({
+      data: {
+        id: "123",
+      },
+    });
+
+    type test = Expect<Equal<typeof result, string>>;
+
+    expect(result).toBe("123");
+  });
+
+  it("Should error when anything else is passed in", () => {
+    expect(() => parseValue("123")).toThrow("Parsing error!");
+    expect(() => parseValue(123)).toThrow("Parsing error!");
+  });
+});
+
+describe("parseValueAgain", () => {
+  it("Should handle a { data: { id: string } }", () => {
+    const result = parseValueAgain({
+      data: {
+        id: "123",
+      },
+    });
+
+    type test = Expect<Equal<typeof result, string>>;
+
+    expect(result).toBe("123");
+  });
+
+  it("Should error when anything else is passed in", () => {
+    expect(() => parseValueAgain("123")).toThrow("Parsing error!");
+    expect(() => parseValueAgain(123)).toThrow("Parsing error!");
+  });
 });
